# fix(services/alert): Improve JSON serialization performance