… services/alert/dao_easyjson.go
… services/storage/version_easyjson.go
… vendor/github.com/mailru/easyjson/bootstrap/bootstrap.go
… vendor/github.com/mailru/easyjson/easyjson/main.go
… vendor/github.com/mailru/easyjson/gen/decoder.go
… vendor/github.com/mailru/easyjson/gen/encoder.go
… vendor/github.com/mailru/easyjson/gen/generator.go
… vendor/github.com/mailru/easyjson/helpers.go
… vendor/github.com/mailru/easyjson/jlexer/lexer.go
… vendor/github.com/mailru/easyjson/jwriter/writer.go
… vendor/github.com/mailru/easyjson/parser/parser.go
… vendor/github.com/mailru/easyjson/parser/parser_unix.go
… vendor/github.com/mailru/easyjson/parser/parser_windows.go
… vendor/github.com/mailru/easyjson/raw.go