Uses mailru/easyjson to reduce the CPU and memory cost of serializing the `TopicState` and `EventState` structs. 

Benchmarks demonstrating the improvements to the `MarshalBinary` and `UnmarshalBinary` APIs:


```
name                                               old time/op    new time/op     delta
TopicState_MarshalBinary/1_event_state(s)-8          5.11µs ± 0%     1.91µs ± 2%   -62.66%  (p=0.008 n=5+5)
TopicState_MarshalBinary/100_event_state(s)-8         326µs ± 1%      104µs ± 0%   -68.14%  (p=0.008 n=5+5)
TopicState_MarshalBinary/30000_event_state(s)-8       111ms ± 2%       39ms ± 2%   -64.70%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/1_event_state(s)-8        9.60µs ± 0%     2.88µs ± 0%   -70.01%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/100_event_state(s)-8       629µs ± 0%      208µs ± 1%   -66.99%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/30000_event_state(s)-8     193ms ± 1%       68ms ± 1%   -64.75%  (p=0.008 n=5+5)

name                                               old speed      new speed       delta
TopicState_MarshalBinary/1_event_state(s)-8        57.3MB/s ± 0%  153.4MB/s ± 2%  +167.78%  (p=0.008 n=5+5)
TopicState_MarshalBinary/100_event_state(s)-8      73.3MB/s ± 1%  230.1MB/s ± 0%  +213.88%  (p=0.008 n=5+5)
TopicState_MarshalBinary/30000_event_state(s)-8    65.6MB/s ± 2%  185.7MB/s ± 2%  +183.27%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/1_event_state(s)-8      30.5MB/s ± 0%  101.8MB/s ± 0%  +233.39%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/100_event_state(s)-8    38.0MB/s ± 0%  115.1MB/s ± 1%  +202.98%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/30000_event_state(s)-8  37.7MB/s ± 1%  107.0MB/s ± 1%  +183.68%  (p=0.008 n=5+5)

name                                               old alloc/op   new alloc/op    delta
TopicState_MarshalBinary/1_event_state(s)-8          2.24kB ± 0%     1.84kB ± 0%   -17.86%  (p=0.008 n=5+5)
TopicState_MarshalBinary/100_event_state(s)-8         179kB ± 0%       56kB ± 0%   -68.77%  (p=0.008 n=5+5)
TopicState_MarshalBinary/30000_event_state(s)-8      48.4MB ± 0%     16.4MB ± 0%   -66.07%  (p=0.016 n=4+5)
TopicState_UnmarshalBinary/1_event_state(s)-8        3.10kB ± 0%     1.47kB ± 0%   -52.45%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/100_event_state(s)-8       150kB ± 0%       72kB ± 0%   -51.63%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/30000_event_state(s)-8    47.6MB ± 0%     26.9MB ± 0%   -43.41%  (p=0.008 n=5+5)

name                                               old allocs/op  new allocs/op   delta
TopicState_MarshalBinary/1_event_state(s)-8            22.0 ± 0%       10.0 ± 0%   -54.55%  (p=0.008 n=5+5)
TopicState_MarshalBinary/100_event_state(s)-8           526 ± 0%        121 ± 0%   -77.00%  (p=0.008 n=5+5)
TopicState_MarshalBinary/30000_event_state(s)-8        150k ± 0%        31k ± 0%   -79.67%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/1_event_state(s)-8          33.0 ± 0%       10.0 ± 0%   -69.70%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/100_event_state(s)-8         839 ± 0%        415 ± 0%   -50.54%  (p=0.008 n=5+5)
TopicState_UnmarshalBinary/30000_event_state(s)-8      241k ± 0%       121k ± 0%   -49.76%  (p=0.008 n=5+5)
```

Fixes #1971

###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [ ] CHANGELOG.md updated