@nathanielc 2018-06-20
Looks straightforward and the performance gains are significant. Is there a go generate comment anywhere? Other than that I think this is good to go.