@stuartcarnie 2018-06-21
I'll add the `go generate` calls